Struct stc_ts_module_loader::ModuleGraph

pub struct ModuleGraph<C, R>where
    C: Comments + Send + Sync,
    R: Resolve,{ /* private fields */ }

Implementation note

This module loader works by

Collect deps recursively (in parallel)

Load all resolved dependencies (in parallel)

Handle all declare module statements (in parallel)

Load all modules again, but this time with all deps resolved. (in parallel)

Build a dependency graph.

Double-loading is required because we have to handle all declare module statements to get all dependencies resolved.

Implementations

impl<C, R> ModuleGraph<C, R>where
C: Comments + Send + Sync,
R: Resolve,

pub fn new(
cm: Arc<SourceMap>,
comments: C,
resolver: R,
parser_config: TsConfig,
target: EsVersion
) -> Self

pub fn comments(&self) -> &C

pub fn load_all(
&self,
entry: &Arc<FileName>
) -> Result<ModuleId, (ModuleId, Error)>

TODO: Fix race condition of errors.

pub fn id_for_declare_module(&self, module_name: &JsWord) -> ModuleId

pub fn path(&self, id: ModuleId) -> Arc<FileName>

pub fn get_circular(&self, id: ModuleId) -> Option<Vec<ModuleId>>

pub fn id(&self, path: &Arc<FileName>) -> ModuleId

pub fn resolve(
&self,
base: &FileName,
specifier: &JsWord
) -> Result<Arc<FileName>, Error>

pub fn clone_module(&self, id: ModuleId) -> Option<Module>

pub fn stmt_count_of(&self, id: ModuleId) -> usize